Move sketch to appropriately named subfolder
The Arduino IDE requires that a sketch be located in a folder of the same name. When opening a file that does not meet this sketch/folder name matching requirement the Arduino IDE presents a dialog:

The file "esp32_I2S_recorder.ino" needs to be inside a sketch folder named "esp32_I2S_recorder".
Create this folder, move the file, and continue?

After clicking "OK" the Arduino IDE currently moves only the file esp32_I2S_recorder.ino to the new folder, leaving behind the other source files. This causes compilation of the sketch to fail:

esp32_SoundRecorder-master\esp32_I2S_recorder\esp32_I2S_recorder.ino:3:17: fatal error: Wav.h: No such file or directory

The alternative solution you might be thinking of is to change the sketch name to match the repository name. This would work for people who clone the repository but when GitHub's popular Clone or download > Download ZIP feature is used to download the contents of a repository the branch/tag/commit name is appended to the folder name, resulting in the sketch and folder names still not matching.